Create compiler adapter factory
