+++ src/plugins/intel_npu/src/compiler_adapter/include/driver_compiler_adapter.hpp
@@ -6,13 +6,10 @@
 
 #pragma once
 
-#include <ze_api.h>
-#include <ze_graph_ext.h>
-
 #include <type_traits>
 #include <utility>
 
-#include "intel_npu/common/npu.hpp"
+#include "intel_npu/common/icompiler_adapter.hpp"
 #include "intel_npu/config/config.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
 #include "intel_npu/utils/zero/zero_init.hpp"
