+++ src/plugins/intel_npu/src/compiler_adapter/src/driver_compiler_adapter.cpp
@@ -4,8 +4,6 @@
 
 #include "driver_compiler_adapter.hpp"
 
-#include <ze_graph_ext.h>
-
 #include <regex>
 #include <string_view>
 
@@ -21,7 +19,6 @@
 #include "intel_npu/utils/zero/zero_utils.hpp"
 #include "ir_serializer.hpp"
 #include "openvino/core/model.hpp"
-#include "ze_graph_ext_wrappers.hpp"
 
 namespace {
