+++ src/plugins/intel_npu/src/plugin/include/plugin.hpp
@@ -52,8 +52,6 @@ public:
                                     const ov::AnyMap& properties) const override;
 
 private:
-    std::unique_ptr<ICompilerAdapter> getCompiler(const Config& config) const;
-
     std::shared_ptr<NPUBackends> _backends;
 
     std::map<std::string, std::string> _config;
