+++ src/plugins/intel_npu/src/plugin/src/plugin.cpp
@@ -7,9 +7,9 @@
 #include <fstream>
 
 #include "compiled_model.hpp"
-#include "npuw/compiled_model.hpp"
-#include "driver_compiler_adapter.hpp"
+#include "compiler_adapter_factory.hpp"
 #include "intel_npu/common/device_helpers.hpp"
+#include "intel_npu/common/icompiler_adapter.hpp"
 #include "intel_npu/common/igraph.hpp"
 #include "intel_npu/common/itt.hpp"
 #include "intel_npu/config/common.hpp"
@@ -17,13 +17,12 @@
 #include "intel_npu/config/npuw.hpp"
 #include "intel_npu/config/runtime.hpp"
 #include "intel_npu/utils/zero/zero_init.hpp"
+#include "npuw/compiled_model.hpp"
 #include "openvino/op/constant.hpp"
 #include "openvino/op/parameter.hpp"
 #include "openvino/runtime/intel_npu/properties.hpp"
 #include "openvino/runtime/properties.hpp"
-#include "plugin_compiler_adapter.hpp"
 #include "remote_context.hpp"
-#include "zero_backend.hpp"
 
 using namespace intel_npu;
 
@@ -699,8 +698,9 @@ std::shared_ptr<ov::ICompiledModel> Plugin::compile_model(const std::shared_ptr<
         }
     }
 
-    auto original_model = model->clone();
-    auto compiler = getCompiler(localConfig);
+    auto originalModel = model->clone();
+    CompilerAdapterFactory compilerAdapterFactory;
+    auto compiler = compilerAdapterFactory.getCompiler(_backends->getIEngineBackend(), localConfig);
 
     OV_ITT_TASK_NEXT(PLUGIN_COMPILE_MODEL, "compile");
     std::shared_ptr<intel_npu::IGraph> graph;
@@ -716,7 +716,7 @@ std::shared_ptr<ov::ICompiledModel> Plugin::compile_model(const std::shared_ptr<
 
     std::shared_ptr<ov::ICompiledModel> compiledModel;
     try {
-        compiledModel = std::make_shared<CompiledModel>(original_model, shared_from_this(), device, graph, localConfig);
+        compiledModel = std::make_shared<CompiledModel>(originalModel, shared_from_this(), device, graph, localConfig);
     } catch (const std::exception& ex) {
         OPENVINO_THROW(ex.what());
     } catch (...) {
@@ -772,7 +772,8 @@ std::shared_ptr<ov::ICompiledModel> Plugin::import_model(std::istream& stream, c
     std::shared_ptr<ov::ICompiledModel> compiledModel;
 
     try {
-        auto compiler = getCompiler(localConfig);
+        CompilerAdapterFactory compilerAdapterFactory;
+        auto compiler = compilerAdapterFactory.getCompiler(_backends->getIEngineBackend(), localConfig);
 
         auto graphSize = getFileSize(stream);
 
@@ -821,7 +822,8 @@ ov::SupportedOpsMap Plugin::query_model(const std::shared_ptr<const ov::Model>& 
     const auto platform = _backends->getCompilationPlatform(localConfig.get<PLATFORM>(), localConfig.get<DEVICE_ID>());
     localConfig.update({{ov::intel_npu::platform.name(), platform}});
 
-    auto compiler = getCompiler(localConfig);
+    CompilerAdapterFactory compilerAdapterFactory;
+    auto compiler = compilerAdapterFactory.getCompiler(_backends->getIEngineBackend(), localConfig);
     ov::SupportedOpsMap supportedOpsMap;
     try {
         supportedOpsMap = compiler->query(model, localConfig);
@@ -832,40 +834,6 @@ ov::SupportedOpsMap Plugin::query_model(const std::shared_ptr<const ov::Model>& 
     }
 
     return supportedOpsMap;
-}
-
-std::unique_ptr<ICompilerAdapter> Plugin::getCompiler(const Config& config) const {
-    auto compilerType = config.get<COMPILER_TYPE>();
-    _logger.debug("performing createCompiler");
-
-    switch (compilerType) {
-    case ov::intel_npu::CompilerType::MLIR: {
-        if (_backends->getBackendName() != "LEVEL0") {
-            return std::make_unique<PluginCompilerAdapter>(nullptr);
-        }
-
-        auto zeroBackend = std::dynamic_pointer_cast<ZeroEngineBackend>(_backends->getIEngineBackend()._ptr);
-        if (zeroBackend == nullptr) {
-            return std::make_unique<PluginCompilerAdapter>(nullptr);
-        }
-
-        return std::make_unique<PluginCompilerAdapter>(zeroBackend->getInitStruct());
-    }
-    case ov::intel_npu::CompilerType::DRIVER: {
-        if (_backends->getBackendName() != "LEVEL0") {
-            OPENVINO_THROW("NPU Compiler Adapter must be used with LEVEL0 backend");
-        }
-
-        auto zeroBackend = std::dynamic_pointer_cast<ZeroEngineBackend>(_backends->getIEngineBackend()._ptr);
-        if (!zeroBackend) {
-            OPENVINO_THROW("Failed to cast zeroBackend, zeroBackend is a nullptr");
-        }
-
-        return std::make_unique<DriverCompilerAdapter>(zeroBackend->getInitStruct());
-    }
-    default:
-        OPENVINO_THROW("Invalid NPU_COMPILER_TYPE");
-    }
 }
 
 std::atomic<int> Plugin::_compiledModelLoadCounter{1};
